kvm: Fixed removal of hosts from certsmap when running certificate auto-renewal (#4156) (#117)
When a host connects to a management server, the host IP address and the certificate are stored in memory on the management server. This mapping is checked periodically to determine if any certificates are due to expire.

Before a certificate is renewed, a few checks are done to determine if the host is connected to the management server by fetching the host record from the database. The problem here is if the wrong record is fetched, the host is not checked for renewal.

This PR improves the host record fetch from the database by looking only at hosts that are not removed.

Fixes: #4129
(cherry picked from commit 7b881517b716e605b8436b3db964d7f83ad0d714)
Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>

Co-authored-by: Spaceman1984 <49917670+Spaceman1984@users.noreply.github.com>