api: Added information about deviceid 0 to attach root volume to VM (#6518)
This is in the workflow of attaching and detaching a root volume of a VM. We can detach the root volume from a VM and Cloudstack marks it as datadisk. We can reattach the volume to make it root volume again. During attach Device ID 0 has to be passed so that the volume will be converted back to ROOT again

Added a small information in the DeviceID section in Attach volume form.