utils: fix human-readable parsing failures
When a API response parameter which is expected by HumanReadableJson utility is not returning a number or Long value it results in API failure.
This change makes API return the value as it is when there is a NumberFormatException with one of the API response param which is listed in HumanReadableJson.

Signed-off-by: Abhishek Kumar <abhishek.mrt22@gmail.com>