KVM: revert libvirtd config and retry if fail to add a host
This PR fixes #6716

If a KVM host has been added to cloudstack environment before, then we reset agent.properties
to original file (./agent/conf/agent.properties), and add the host to cloudstack again, we
will face the following error:

2023-01-11 20:54:03,734 DEBUG [c.c.h.k.d.LibvirtServerDiscoverer] (qtp1647766367-13:ctx-17f9aff7 ctx-47c881cb) (logid:49819f17)  can't setup agent, due to com.cloud.utils.exception.CloudRuntimeException: Failed to setup keystore on the KVM host: 10.0.32.229 - Failed to setup keystore on the KVM host: 10.0.32.229
com.cloud.utils.exception.CloudRuntimeException: Failed to setup keystore on the KVM host: 10.0.32.229
        at com.cloud.hypervisor.kvm.discoverer.LibvirtServerDiscoverer.setupAgentSecurity(LibvirtServerDiscoverer.java:178)
        at com.cloud.hypervisor.kvm.discoverer.LibvirtServerDiscoverer.find(LibvirtServerDiscoverer.java:320)

This is because the libvirtd has been configured to listen on LTS port when the host
was added to cloudstack before. When try to add host again, cloudstack will generate a new
keystore passphrase which is not same as before, and wrong password for existing keystore.

The issue can be fixed by
(1) revert libvirtd config to not use TLS
(2) remove all existing keystore/certificate
(3) generate new keystore/certificate
(4) setup agent, etc