Removing use of SmartConnectNoSSL and adding custom SSL context
This PR removes the dependency on the deprecated method SmartConnectNoSSL from the pyVim.connect module. The pyVmomi library removed this method in newer versions, leading to errors when trying to import it.

Instead of using SmartConnectNoSSL, I have modified the code to use SmartConnect with a custom SSL context when SSL verification is disabled. Python has an inbuilt method to create an SSL context that does not verify the authenticity of the SSL certificate, ssl._create_unverified_context(). This method is used to create a context which is passed to SmartConnect.

The following changes were made to the code:

Import of ssl added at the top of the file.
SmartConnectNoSSL was removed from the imports from the pyVim.connect module.
Modified the block of code where SmartConnect or SmartConnectNoSSL was used to create a service instance. It now creates an unverified SSL context if the disable_ssl_verification option is set and Python supports creating unverified SSL contexts. This context is then passed to SmartConnect.
This PR helps to maintain compatibility with newer versions of pyVmomi and avoids using deprecated methods.

Finally, it should be noted that disabling SSL verification can lead to potential security issues. This option should be used with care and preferably in a testing or development environment.