Veeam: import powershell module which is available since Veeam 11
The following comes from Veeam release note
```
Mods: please move to the PowerShell forum.

Yes, this is known change to a PowerShell module from a snap-in, it was made during a beta stage for v11. The requirements for PowerShell 5.1 (up from 2) are covered in the Release Notes, and the details about the PowerShell module are covered in the "What's New" documentation.

Details are here:

PowerShell
• PowerShell module — By popular demand, we switched from the PowerShell snap-in to the PowerShell module, which can be used on any machine with the backup console installed. We also no longer require PowerShell 2.0 installed on the backup server, which is something many customers had problems with.
• New PowerShell cmdlet — Version 11 adds 184 new cmdlets for both newly added functionality and expanded coverage of the existing features with a particular focus on restore functionality.
```

This fixes the error message when assign VM to backup offering
```
2023-10-25 12:35:11,032 DEBUG [o.a.c.b.v.VeeamClient] (API-Job-Executor-5:ctx-5f068dbb job-42 ctx-3fbfe1a2) (logid:5840110c) Veeam response for PowerShell commands [PowerShell Add-PSSnapin VeeamPSSnapin;$Job = Get-VBRJob -name "template_job_zone1_default";$Job.GetBackupTargetRepository() ^| select Name ^| Format-List] is: [^M
^M
Name : Default Backup Repository^M
^M
^M
^M
Add-PSSnapin : No snap-ins have been registered for Windows PowerShell version 5.^M
At line:1 char:1^M
+ Add-PSSnapin VeeamPSSnapin;$Job = Get-VBRJob -name template_job_zone1 ...^M
+ ~~~~~~~~~~~~~~~~~~~~~~~~~~^M
    + CategoryInfo          : InvalidArgument: (VeeamPSSnapin:String) [Add-PSSnapin], PSArgumentException^M
    + FullyQualifiedErrorId : AddPSSnapInRead,Microsoft.PowerShell.Commands.AddPSSnapinCommand^M
 ^M
].
```