api,server: allow listing diskofferings with suitability for vm
VMs deployed with specific resources may not support creating volumes
with all disk offerings. This change enables listing suitable disk
offerings for a VM.
A new parameter - virtualmachineid has been added to the
listDiskOfferings API.
When this parameter is passed it returns response parameter -
suitableforvirtualmachine for each offering in the response list.

Signed-off-by: Abhishek Kumar <abhishek.mrt22@gmail.com>