ui: change reporting link to Github Discussions
Many users are using the footer link to open questions about CloudStack
that are usually discussed on the users@ mailing list. This fixes that
behaviour by diverting them to Github Discussions which are linked with
the user@ ML, smart users can still report actual bugs/issues via the
issues tab.

Signed-off-by: Rohit Yadav <rohit.yadav@shapeblue.com>